Move `termutil`.
